Flytter ut uthenting av andelshistorikk i egen service for å unngå circular dependencies då den var avhengig av VurderingService som er avhengig av 1000 andre services
