+++ src/main/kotlin/no/nav/familie/ef/sak/behandlingsflyt/steg/BeregnYtelseSteg.kt
@@ -17,6 +17,7 @@ import no.nav.familie.ef.sak.infrastruktur.exception.brukerfeilHvis
 import no.nav.familie.ef.sak.infrastruktur.exception.feilHvis
 import no.nav.familie.ef.sak.simulering.SimuleringService
 import no.nav.familie.ef.sak.tilbakekreving.TilbakekrevingService
+import no.nav.familie.ef.sak.tilkjentytelse.AndelsHistorikkService
 import no.nav.familie.ef.sak.tilkjentytelse.TilkjentYtelseService
 import no.nav.familie.ef.sak.tilkjentytelse.domain.AndelTilkjentYtelse
 import no.nav.familie.ef.sak.tilkjentytelse.domain.TilkjentYtelse
@@ -46,6 +47,7 @@ import kotlin.reflect.KClass
 @Service
 class BeregnYtelseSteg(
     private val tilkjentYtelseService: TilkjentYtelseService,
+    private val andelsHistorikkService: AndelsHistorikkService,
     private val beregningService: BeregningService,
     private val beregningBarnetilsynService: BeregningBarnetilsynService,
     private val beregningSkolepengerService: BeregningSkolepengerService,
@@ -122,7 +124,7 @@ class BeregnYtelseSteg(
     }
 
     private fun validerStartTidEtterSanksjon(vedtakFom: YearMonth, behandling: Saksbehandling) {
-        val nyesteSanksjonsperiode = tilkjentYtelseService.hentHistorikk(behandling.fagsakId, null)
+        val nyesteSanksjonsperiode = andelsHistorikkService.hentHistorikk(behandling.fagsakId, null)
             .lastOrNull { it.periodeType == VedtaksperiodeType.SANKSJON }
         nyesteSanksjonsperiode?.andel?.stønadFra?.let { sanksjonsdato ->
             feilHvis(sanksjonsdato >= vedtakFom.atDay(1)) {
